Add charname command
Adds a charname command to change the name of the command issueing
user's character to the option supplied string. Because the user needed
to be found by the id of the command issueing user rather than a user
form an option db.findOneUser has been changed to accept the userID
instead of the interaction, and commands also using the function have
been updated accordingly.