Add map and zones
Adds a zone class which is called by a zones index to return a zone base
on a zone object in the index. Two new db module methods are added along
with log methods to create and find zones in the database which use the
zones index. A map index has been added to assign zones to coordinates
for the upcomming movement system.