Add spawn command and refactor
Adds a spawn command that displays RPG new player info then teleports
the user's char to the middle of the map. Adds a new userChecks module
used by all commands that need to check if a user already exists in the
DB or if a char has already spawned. Commands have been refactored to
use the module and the db module had been refactored to comply.