Add combat module
Adds a new combat module that will handle all combat calculations and
return it's data to the command response. Currently contains a calcToHit
helper function that will be called by the combat methods to calculate
the toHit and toCrit numbers based on weapon skill to be compared to a
hit roll.