Add helpers to combat module
Adds 2 helper functions to the combat module. calcInitiative, which
calculates and returns an int for initiative based on armorSkill and
an armor penatly equal to maxArmorValue to be compared to another
initiatve to determine turn order. rollHit, which returns a random int
between 1 and 100. The consts from calcToHit have also moved out to be
global and 2 new ones added. calToHit has been corrected for typos.