Add responses module
Adds a new responses module that contains responses for select commands,
currently all the non-testing commands for RPG. Commands will call a
responses method passing in the needed data and the method will in turn
call the respond method that calls the interaction.reply method in a try
catch. This will greatly increase reability and shorten the command
files as well as provide reusability.